% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-hit-100.R
\name{add_country_hit_100}
\alias{add_country_hit_100}
\title{Add a column that indicates when a country hit 100}
\usage{
add_country_hit_100(covid_data)
}
\arguments{
\item{covid_data}{covid19 data with cumulative cases for each country}
}
\value{

}
\description{
Requires that country is called \code{country_region}, date is called \code{date},
cumulative cases are called \code{cumulative_cases}.
}
\examples{
# covid \%>\%
}
