% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutting-room.R
\name{add_country_label}
\alias{add_country_label}
\title{Add a country label for plotting}
\usage{
add_country_label(data, countries = unique(data$country_region))
}
\arguments{
\item{data}{covid19 data}

\item{countries}{default adds all countries. Otherwise, they are
specified with character vector.}
}
\value{
data.frame with added columns of selected countries
}
\description{
This adds the appropriate lables for each country, to assist plotting
with geom_label
}
