% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-country-row-id.R
\name{add_country_row_id}
\alias{add_country_row_id}
\title{Add row ID for each country from start date}
\usage{
add_country_row_id(covid_data)
}
\arguments{
\item{covid_data}{covid19 data}
}
\value{
adds row ID column, \code{rnum}, for each country
}
\description{
Add row ID for each country from start date
}
