% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-over-limit.R
\name{add_days_since_limit}
\alias{add_days_since_limit}
\title{Add a column of the days since some limit of cumulative cases is reached}
\usage{
add_days_since_limit(covid_data, limit = 100)
}
\arguments{
\item{covid_data}{covid19 data}

\item{limit}{limit for flagging}
}
\value{
data.frame with extra column, "days_since_limit", which contains
the number of days since the cumulative cases reached 100.
}
\description{
Add a column of the days since some limit of cumulative cases is reached
}
