% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-instant-reproduction.R
\name{add_instant_reproduction}
\alias{add_instant_reproduction}
\title{Add instant reproduction number measures to provided covid19 data}
\usage{
add_instant_reproduction(covid_data)
}
\arguments{
\item{covid_data}{covid19 data}
}
\value{
data.frame with covid19 columns plus output of \code{EpiEstim::estimate_R}
and columns on estimated mean R values and quantiles etc.
}
\description{
Add instant reproduction number measures to provided covid19 data
}
