% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-weekly-avg.R
\name{add_weekly_avg}
\alias{add_weekly_avg}
\title{Add Weekly average summary, for each country}
\usage{
add_weekly_avg(covid, ...)
}
\arguments{
\item{covid}{covid19 data}

\item{...}{variables to compute weekly average of}
}
\value{
data frame with weekly average computed
}
\description{
Add Weekly average summary, for each country
}
\examples{
\dontrun{
covid <- covid_latest()
add_weekly_avg(covid, cases)
add_weekly_avg(covid, deaths)
add_weekly_avg(covid, cases, deaths)
}
}
