% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-instant-reproduction.R
\name{augment_estimate_repro}
\alias{augment_estimate_repro}
\title{Tidies output of \code{estimate_repro_all}}
\usage{
augment_estimate_repro(covid_data_estimated)
}
\arguments{
\item{covid_data_estimated}{covid19 data estimated with \code{estimate_repro_all}}
}
\value{

}
\description{
Tidies output of \code{estimate_repro_all}
}
