% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-instant-reproduction.R
\name{country_repro_errors}
\alias{country_repro_errors}
\title{Extract (any) errors from \code{estimate_repro_all}}
\usage{
country_repro_errors(covid_data_estimated)
}
\arguments{
\item{covid_data_estimated}{data from \code{estimate_repro_all}}
}
\value{
vector of error messages
}
\description{
Extract (any) errors from \code{estimate_repro_all}
}
