% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-instant-reproduction.R
\name{covid-estimate-repro}
\alias{covid-estimate-repro}
\alias{estimate_repro_all}
\title{Fits instant reproduction for each country}
\usage{
estimate_repro_all(
  covid_data,
  estimate_method = "parametric_si",
  si_sample = NULL,
  config = list(mean_si = 4.7, std_si = 2.9)
)
}
\arguments{
\item{covid_data}{covid19 data}

\item{estimate_method}{default is "parametric_si"}

\item{si_sample}{SI samples. Default is NULL.}

\item{config}{named list of values to pass to EpiEstim::make_config. Default
is \code{mean_si = 4.7}, and \code{std_si = 2.9}.}
}
\value{
data.frame with several list columns containing various output of
\code{EpiEstim::estimate_R}.
}
\description{
Fits instant reproduction for each country
}
