% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{covid_change_point}
\alias{covid_change_point}
\title{Title}
\usage{
covid_change_point(covid_data, ...)
}
\arguments{
\item{...}{params to pass to \code{changepoint::cpt.meanvar}}
}
\value{

}
\description{
Title
}
