% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-covid-cdc.R
\name{covid_ecdc}
\alias{covid_ecdc}
\title{Download raw COVID19 data from European CDC}
\usage{
covid_ecdc(date)
}
\arguments{
\item{date}{in format "YYYY-MM-DD". So, "2020-01-01" is 1st Jan, 2020}
}
\value{
data.frame
}
\description{
This function will download raw COVID19 data, given a date.
It will error if you enter a date that is later than the current date
on the CDC site. See \code{latest_covid()} for a function that handles
the full details of this.
}
