% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-instant-reproduction.R
\name{covid_estimate_repro}
\alias{covid_estimate_repro}
\title{Fit EpiEstim::estimate_R to covid19 data}
\usage{
covid_estimate_repro(covid_data_prepared, estimate_method = "parametric_si")
}
\arguments{
\item{covid_data_prepared}{data is in the format provided by
\code{covid_prepare_estimate_repro}, with columns "dates" and "I", where "dates"
contains data of class "Date", and "I" contains cases.}

\item{estimate_method}{default is "parametric_si"}
}
\value{
estimated reproductive number
}
\description{
This fits EpiEstim::estimate_R, with config values \code{mean_si = 5.0} and
\code{std_si = 3.0}. It assumes that the data is in the format provided by
\code{covid_prepare_estimate_repro}.
}
