% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-covid-data.R
\name{covid_filter_countres}
\alias{covid_filter_countres}
\title{Filter covid countries to provided countries}
\usage{
covid_filter_countres(data, countries)
}
\arguments{
\item{data}{covid19 data}

\item{countries}{vector of countries}
}
\value{
dataset filtered to existing countries
}
\description{
Filter covid countries to provided countries
}
