% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{covid_highlight_australia_plus}
\alias{covid_highlight_australia_plus}
\title{Add alpha column which will highlight australia foremost, then ohters}
\usage{
covid_highlight_australia_plus(data, plus = "Canada")
}
\arguments{
\item{data}{covid19 data}

\item{plus}{additional countries to highlight. Defaults to just Canada.}
}
\value{
data with added column "alpha" with extra info
}
\description{
Add alpha column which will highlight australia foremost, then ohters
}
