% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-covid-jhu.R
\name{covid_jhu_clean_names}
\alias{covid_jhu_clean_names}
\title{clean variable names of covid19 raw data (from \code{covid_read})}
\usage{
covid_jhu_clean_names(x)
}
\arguments{
\item{x}{covid19 raw data}
}
\value{
data.frame with tidied names
}
\description{
clean variable names of covid19 raw data (from \code{covid_read})
}
