% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-covid-data.R
\name{covid_pivot_longer}
\alias{covid_pivot_longer}
\title{pivots the covid19 data into longer form}
\usage{
covid_pivot_longer(x)
}
\arguments{
\item{x}{data that has been imported and cleaned}
}
\value{
longer form pivoted data
}
\description{
pivots the covid19 data into longer form
}
