% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-instant-reproduction.R
\name{covid_prepare_estimate_repro}
\alias{covid_prepare_estimate_repro}
\title{Prepare covid19 data for estimating instant repro number}
\usage{
covid_prepare_estimate_repro(covid_data)
}
\arguments{
\item{covid_data}{covid19 data}
}
\value{
dataframe with columns "dates" and "I", ready for
\code{EpiEstim::estimate_R}
}
\description{
Prepare covid19 data for estimating instant repro number
}
