% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-covid-data.R
\name{covid_pull_data}
\alias{covid_pull_data}
\title{Pulls and tidies the latest covid19 data}
\usage{
covid_pull_data()
}
\value{
covid19 data with columns..
}
\description{
Pulls and tidies the latest covid19 data
}
