% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed-time-to-double.R
\name{covid_summarise_doubling}
\alias{covid_summarise_doubling}
\title{Calculate the time to double covid19 data}
\usage{
covid_summarise_doubling(covid_data)
}
\arguments{
\item{covid_data}{covid19 data}
}
\value{
data of country region, with \code{days_to_double}, and \code{double_target}
being the "from x number of cases"
}
\description{
Calculate the time to double covid19 data
}
\author{
Miles McBain
}
