% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa-utils.R
\name{diff_days}
\alias{diff_days}
\title{Difference in days between two dates}
\usage{
diff_days(x, y)
}
\arguments{
\item{x}{first date}

\item{y}{second date}
}
\value{
numeric number of days
}
\description{
Difference in days between two dates
}
