% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{effective-repro-incidence}
\alias{effective-repro-incidence}
\alias{gg_effective_repro_incidence_patchwork}
\title{Effective Reproductive pathwork}
\usage{
gg_effective_repro_incidence_patchwork(covid_effective_r, covid_data, country)
}
\description{
Effective Reproductive pathwork
}
