% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{gg_effective_repro_all}
\alias{gg_effective_repro_all}
\alias{effective-repro}
\alias{gg_effective_repro_facet}
\title{Plot effective reproduction number}
\usage{
gg_effective_repro_all(covid_effective_r)

gg_effective_repro_facet(covid_effective_r)
}
\arguments{
\item{covid_effective_r}{covid19 data with estimated effective R}
}
\value{
ggplot2 plot
}
\description{
Plot effective reproduction number
}
