% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{extract_change_point_mean_var}
\alias{extract_change_point_mean_var}
\title{Extract changepoint}
\usage{
extract_change_point_mean_var(data, x = "cases", ...)
}
\arguments{
\item{data}{data.frame}

\item{x}{variable name (as character) of cases. Default is "cases".}

\item{...}{params to pass to \code{changepoint::cpt.meanvar}}
}
\value{
changepoint vars
}
\description{
Extract changepoint
}
