% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{extract_first_np_change_point}
\alias{extract_first_np_change_point}
\title{Extract non-parametric changepoint}
\usage{
extract_first_np_change_point(data, x = "cases", ...)
}
\arguments{
\item{data}{data.frame}

\item{x}{variable name (as character) of cases. Default is "cases".}

\item{...}{params to pass to \code{changepoint.np::cpt.np}}
}
\value{
changepoint vars
}
\description{
Extract non-parametric changepoint
}
