% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-over-100.R
\name{filter_country_over_100}
\alias{filter_country_over_100}
\title{Filter data to only include countries that have more than 100 cases}
\usage{
filter_country_over_100(covid_data)
}
\arguments{
\item{covid_data}{covid19 data}
}
\value{
data.frame with covid data that has more than 100 cases
}
\description{
Filter data to only include countries that have more than 100 cases
}
