% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-over-limit.R
\name{filter_country_over_limit}
\alias{filter_country_over_limit}
\title{Filter data to only include countries that have more than a certain number
of cases}
\usage{
filter_country_over_limit(covid_data, limit = 100)
}
\arguments{
\item{covid_data}{covid19 data}

\item{limit}{number of cases - default over 100}
}
\value{
data.frame with covid data that has more than 100 cases
}
\description{
Filter data to only include countries that have more than a certain number
of cases
}
