% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{geom_covid_country_label}
\alias{geom_covid_country_label}
\title{Adds nice country labels to ggplot, based on \code{add_country_label()}}
\usage{
geom_covid_country_label()
}
\value{
ggplot label added to ggplot
}
\description{
Adds nice country labels to ggplot, based on \code{add_country_label()}
}
