% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{gg_covid_cases}
\alias{gg_covid_cases}
\alias{gg-covid}
\alias{gg_covid_cumulative_cases}
\title{misc-ggplot funs for covid19}
\usage{
gg_covid_cases(covid_data)

gg_covid_cumulative_cases(covid_data)
}
\arguments{
\item{covid_data}{covid19 data}
}
\value{
ggplot image
}
\description{
misc-ggplot funs for covid19
}
