% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{gg_covid_cumulative_cases_deaths_exceed_limit}
\alias{gg_covid_cumulative_cases_deaths_exceed_limit}
\title{Base plot for creating cumulative cases vs deaths of covid19 data}
\usage{
gg_covid_cumulative_cases_deaths_exceed_limit(
  covid_data_limit,
  limit = NULL,
  highlight = "Australia"
)
}
\arguments{
\item{covid_data_limit}{\itemize{
\item covid19 with added limit
(from \code{\link[=add_days_since_limit]{add_days_since_limit()}})
}}

\item{limit}{the number of days since reached a limit (added for
titling graphic). Default is 100.}

\item{highlight}{the name of the country to highlight, default is Australia.}
}
\value{
ggplot plot
}
\description{
Base plot for creating cumulative cases vs deaths of covid19 data
}
