% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{gg_covid_cumulative_exceed_100}
\alias{gg_covid_cumulative_exceed_100}
\title{Base plot for creating cumulative cases of covid19 data}
\usage{
gg_covid_cumulative_exceed_100(covid_data)
}
\arguments{
\item{covid_data}{\itemize{
\item covid19 data.frame
}}
}
\value{
ggplot plot
}
\description{
Base plot for creating cumulative cases of covid19 data
}
