% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{gg_covid_incidence_exceed_limit}
\alias{gg_covid_incidence_exceed_limit}
\title{Base plot for creating incident cases chart of covid19 data}
\usage{
gg_covid_incidence_exceed_limit(
  covid_data_limit,
  limit = NULL,
  highlight = "Australia"
)
}
\arguments{
\item{covid_data_limit}{\itemize{
\item covid19 with added limit
(from \code{\link[=add_days_since_limit]{add_days_since_limit()}})
}}

\item{limit}{the number of days since reached a limit (added for
titling graphic). Default is 100.}

\item{highlight}{the name of the country to highlight, default is Australia.}
}
\value{
ggplot plot
}
\description{
Base plot for creating incident cases chart of covid19 data
}
\examples{
\dontrun{
 covid_data_since <- covid_data %>%
   add_days_since_limit(limit = 100) %>%
   dplyr::filter(days_since_limit >= 0) %>%
   dplyr::filter(country_region %in% c("Australia", "New Zealand"))

gg_covid_cumulative_exceed_limit(covid_data_since)
}
}
