% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helpers.R
\name{gg_effective_repro_incidence_patchwork}
\alias{gg_effective_repro_incidence_patchwork}
\title{Effective Reproductive pathwork}
\usage{
gg_effective_repro_incidence_patchwork(covid_effective_r, covid_data, country)
}
\arguments{
\item{covid_effective_r}{covid19 data with effective repro added with
\code{covid_estimate_repro()}.}

\item{covid_data}{covid19 data}

\item{country}{country to display}
}
\description{
Effective Reproductive pathwork
}
