% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when-all-gt-zero.R
\name{when_all_gt_zero}
\alias{when_all_gt_zero}
\alias{gt-zero}
\alias{where_all_gt_zero}
\title{Determine when all numbers in a vector are greater than zero}
\usage{
when_all_gt_zero(x)

where_all_gt_zero(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical vector - TRUE when all numbers start to be GT zero
}
\description{
Determine when all numbers in a vector are greater than zero
}
\examples{
when_all_gt_zero(c(0, 1, 0, 1, 1, 2, 3))
when_all_gt_zero(c(0, 0, 0, 0))
when_all_gt_zero(c(1, 0, 1, 1, 0, 2, 3, 4))
where_all_gt_zero(c(1, 0, 1, 1, 0, 2, 3, 4))
}
