% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-china-data.R
\name{patch_china_data}
\alias{patch_china_data}
\title{Patch the china case records}
\usage{
patch_china_data(data)
}
\arguments{
\item{data}{data pulled by}
}
\value{
data.frame wtih adjusted cases
}
\description{
In Mid February, China changed their case definition, from
RT-PCR-swab-positive to test-positive or clinical signs of viral pneumonia.
This led to a large spike not representative of cases. some further thoughts: \url{https://www.worldometers.info/coronavirus/how-to-interpret-feb-12-case-surge/}
This changes the cases from date: 13/2/2020 to 1820 cases (from 15141),
and on date: 4/2/2020 to 1995 cases (from 3237).
}
\note{
internal
}
