% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-covid-cdc.R
\name{pluck_latest_ecdc}
\alias{pluck_latest_ecdc}
\title{Pluck the latest ecdc data}
\usage{
pluck_latest_ecdc(data)
}
\arguments{
\item{data}{covid19 list of latest data pulled from \code{try_ecdc}.}
}
\value{
single data.frame
}
\description{
For internal use within covid19 extraction data
}
\note{
internal
}
