% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-over-100.R
\name{summarise_country_over_100}
\alias{summarise_country_over_100}
\title{Create a summary for each country, and a flag of whether they are over 100}
\usage{
summarise_country_over_100(covid_data)
}
\arguments{
\item{covid_data}{covid19 data}
}
\value{
data.frame with two columns, \code{country_region} and \code{over_100}
}
\description{
Create a summary for each country, and a flag of whether they are over 100
}
