% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-over-limit.R
\name{summarise_country_over_limit}
\alias{summarise_country_over_limit}
\title{Create a summary for each country, and a flag of whether they are over limit}
\usage{
summarise_country_over_limit(covid_data, limit = 100)
}
\arguments{
\item{covid_data}{covid19 data}

\item{limit}{limit}
}
\value{
data.frame with two columns, \code{country_region} and \code{over_limit}
}
\description{
Create a summary for each country, and a flag of whether they are over limit
}
