% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-global-incidence.R
\name{summarise_covid_global}
\alias{summarise_covid_global}
\title{Summarise covid19 data to cases over time globally}
\usage{
summarise_covid_global(covid_data)
}
\arguments{
\item{covid_data}{covid19 data}
}
\value{
data.frame with covid19 cases, columns called
\code{date}, \code{cases}, and \code{cumulative_cases}.
}
\description{
Requires that incidence cases are called \code{cases}, country is called
\code{country_region}, and date is called \code{date}.
}
