% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-instant-reproduction.R
\name{tidy_repro_estimate}
\alias{tidy_repro_estimate}
\title{Tidies up output from \code{EpiEstim::estimate_R}}
\usage{
tidy_repro_estimate(covid_estimated_reproduction)
}
\arguments{
\item{covid_estimated_reproduction}{output from \code{EpiEstim::estimate_R}}
}
\value{
returns columns on estiamted isntant reproduction, and quantiles,
as well as dates.
}
\description{
Tidies up output from \code{EpiEstim::estimate_R}
}
