% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-covid-cdc.R
\name{try_covid_yesterday_today}
\alias{try_covid_yesterday_today}
\title{Downloads both yesterday and today's COVID19 data}
\usage{
try_covid_yesterday_today()
}
\value{
list of yesterdays and today's COVID19 data
}
\description{
Assuming CET time zones. For use within \code{latest_covid()}.
}
