% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-covid-cdc.R
\name{try_ecdc}
\alias{try_ecdc}
\title{Downloads both yesterday and today's COVID19 data}
\usage{
try_ecdc(memoise = TRUE)
}
\arguments{
\item{memoise}{do you want to memoise (cache the data?). Default is TRUE}
}
\value{
list of yesterdays and today's COVID19 data
}
\description{
Assuming CET time zones. For use within \code{latest_covid()}.
}
