% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-over-limit.R
\name{add_country_hit_limit}
\alias{add_country_hit_limit}
\title{Add a column that indicates when a country hits a limit of cases}
\usage{
add_country_hit_limit(covid_data, limit = 100)
}
\arguments{
\item{covid_data}{covid19 data with cumulative cases for each country}

\item{limit}{defaults to 100}
}
\value{
data.frame with additional column "hit limit" indicating the day
a country passed that limit
}
\description{
Requires that country is called \code{country_region}, date is called \code{date},
cumulative cases are called \code{cumulative_cases}.
}
\examples{
# covid \%>\%
}
