% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-instant-reproduction.R
\name{covid_estimate_repro}
\alias{covid_estimate_repro}
\title{Fit EpiEstim::estimate_R to covid19 data}
\usage{
covid_estimate_repro(
  covid_data_prepared,
  estimate_method = "parametric_si",
  si_sample = NULL,
  config = list(mean_si = 4.7, std_si = 2.9)
)
}
\arguments{
\item{covid_data_prepared}{data is in the format provided by
\code{covid_prepare_estimate_repro}, with columns "dates" and "I", where "dates"
contains data of class "Date", and "I" contains cases.}

\item{estimate_method}{default is "parametric_si"}

\item{si_sample}{SI samples. Default is NULL.}

\item{config}{named list of values to pass to EpiEstim::make_config. Default
is \code{mean_si = 4.7}, and \code{std_si = 2.9}.}
}
\value{
estimated reproductive number
}
\description{
This fits EpiEstim::estimate_R, with config values \code{mean_si = 4.7} and
\code{std_si = 2.9} from https://doi.org/10.1101/2020.02.03.20019497
It assumes that the data is in the format provided by
\code{covid_prepare_estimate_repro}.
}
