% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high-incidence.R
\name{covid_high_incidence}
\alias{covid_high_incidence}
\title{get the covid_high_incidence data}
\usage{
covid_high_incidence(covid_data)
}
\arguments{
\item{covid_data}{data pulled from \code{covid_pull_data()}}
}
\value{
high incidence data
}
\description{
get the covid_high_incidence data
}
