% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-covid-cdc.R
\name{covid_latest}
\alias{covid_latest}
\title{Pull latest covid19 data from European CDC}
\usage{
covid_latest(patch = TRUE)
}
\arguments{
\item{patch}{logical. Patch China miscounts see \code{\link[=patch_china_data]{patch_china_data()}}.
Default is TRUE.}
}
\value{
data.frame
}
\description{
This pull data from \url{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}
}
