% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-instant-reproduction.R
\name{estimate_repro_all}
\alias{estimate_repro_all}
\title{Fits instnat reproduction for each country}
\usage{
estimate_repro_all(covid_data)
}
\arguments{
\item{covid_data}{covid19 data}
}
\value{
data.frame with several list columns containing various output of
\code{EpiEstim::estimate_R}.
}
\description{
Fits instnat reproduction for each country
}
