% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-instant-reproduction.R
\name{estimate_repro_all}
\alias{estimate_repro_all}
\title{Fits instnat reproduction for each country}
\usage{
estimate_repro_all(
  covid_data,
  estimate_method = "parametric_si",
  si_sample = NULL,
  ...
)
}
\arguments{
\item{covid_data}{covid19 data}

\item{estimate_method}{default is "parametric_si"}

\item{si_sample}{default is NULL - matrix of values related to the Serial
interval, the time from onset of symptoms.}

\item{...}{named list of arguments to pass to \code{config_list}. If no arguments
presented, defaults to \code{mean_si = 5.0}, and \code{std_si = 3.0}}
}
\value{
data.frame with several list columns containing various output of
\code{EpiEstim::estimate_R}.
}
\description{
Fits instnat reproduction for each country
}
