% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{extract_change_point_mean_var}
\alias{extract_change_point_mean_var}
\title{Extract changepoint}
\usage{
extract_change_point_mean_var(data, ...)
}
\arguments{
\item{...}{params to pass to \code{changepoint::cpt.meanvar}}
}
\value{

}
\description{
Extract changepoint
}
