% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutting-room.R
\name{gg_covid_cumulative_exceed_limit}
\alias{gg_covid_cumulative_exceed_limit}
\title{Base plot for creating cumulative cases of covid19 data}
\usage{
gg_covid_cumulative_exceed_limit(
  covid_data,
  limit = 100,
  selected_countries = "Australia"
)
}
\arguments{
\item{covid_data}{\itemize{
\item covid19 data.frame
}}

\item{limit}{limit of cumulative cases}

\item{selected_countries}{\itemize{
\item character vector of countries to plot
}}
}
\value{
ggplot plot
}
\description{
Base plot for creating cumulative cases of covid19 data
}
