% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when-all-gt-zero.R
\name{when_all_gt}
\alias{when_all_gt}
\alias{gt-zero}
\alias{where_all_gt}
\alias{when_all_gte}
\alias{where_all_gte}
\title{Determine when all numbers in a vector are greater than zero}
\usage{
when_all_gt(x, num = 0)

where_all_gt(x, num = 0)

when_all_gte(x, num = 0)

where_all_gte(x, num = 0)
}
\arguments{
\item{x}{numeric vector}

\item{num}{greater than some number, default 0}
}
\value{
logical vector - TRUE when all numbers start to be GT zero
}
\description{
Determine when all numbers in a vector are greater than zero
}
\examples{
when_all_gt(c(0, 1, 0, 1, 1, 2, 3))
when_all_gt(c(0, 0, 0, 0))
when_all_gt(c(1, 0, 1, 1, 0, 2, 3, 4))
where_all_gt(c(1, 0, 1, 1, 0, 2, 3, 4))
}
