% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-china-data.R
\name{patch_china_data}
\alias{patch_china_data}
\title{Patch the china case records}
\usage{
patch_china_data(data)
}
\arguments{
\item{data}{data pulled by \code{try_ecdc} and \code{pluck_latest}.}
}
\value{
data.frame wtih adjusted cases
}
\description{
In Mid February, China changed their case definition, from
RT-PCR-swab-positive to test-positive or clinical signs of viral pneumonia.
This led to a large spike not representative of cases. Numbers sourced from  \url{https://en.wikipedia.org/w/index.php?title=Template:2019–20_coronavirus_pandemic_data/China_medical_cases_by_province&oldid=946973173}.
  This changes the cases as follows date:
}
\details{
\preformatted{- 04/2/2020 to 1995 cases (from 3237),
- 13/2/2020 to 1820 cases (from 15141),
- 15/2/2020 to 1503 cases (from 2538),
- 16/2/2020 to 1121 cases (from 2007),
}
}
\note{
internal
}
